backend-fastapi/requirements.txt
# FastAPI + AD Auth + SMB – Inyatsi File Portal
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12

# AD / LDAP
ldap3>=2.9.1

# SMB file server
smbprotocol>=1.10.0

# Config
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Windows NTFS ACL - required for /api/permissions (install on Windows; skip on Linux)
pywin32>=306; sys_platform == "win32"
